Remove calls of os.Exit from influx cli Run method and fix influx tests
One of the first unit tests in the cli tests called the Run method.
Since the Run method called os.Exit, it reported the unit tests as
succeeded. When parallel is set to 1, this skips _all_ unit tests after
the first one. When parallel is set to a higher value, unit tests run by
other processes still get run.

This changes the Run method to return an error (if one occurred). This
error can then be printed out and a bad exit status can be used to exit
the program from the main program instead.  That causes the unit tests
to run correctly regardless of how many parallel processes are running.

Also added an additional option to the CLI called `IgnoreSignals`. If
this is set to true, then signals are not registered with the process.
Setting signals doesn't really work in unit tests so it's good to ensure
they don't get set in the first place.

In addition to fixing the influx cli tests, this adds a mock client to
the cli test for Use. PR #5183 added a validation for `use` to only be
able to select public databases so `_internal` couldn't be chosen. To
implement this, the `SHOW DATABASES` command was used by the internal
client.

Some of the unit tests in `cli_test.go` don't set the client to
anything. `TestParseCommand_Use` previously didn't, but now it needs to
have a client in the unit test with an empty test server.

+++ cmd/influx/cli/cli.go
@@ -52,6 +52,7 @@ type CommandLine struct {
 	Path             string
 	Compressed       bool
 	Quit             chan struct{}
+	IgnoreSignals    bool // Ignore signals normally caught by this process (used primarily for testing)
 	osSignals        chan os.Signal
 	historyFile      *os.File
 }
@@ -66,9 +67,11 @@ func New(version string) *CommandLine {
 }
 
 // Run executes the CLI
-func (c *CommandLine) Run() {
+func (c *CommandLine) Run() error {
 	// register OS signals for graceful termination
-	signal.Notify(c.osSignals, os.Kill, os.Interrupt, syscall.SIGTERM, syscall.SIGQUIT, syscall.SIGHUP)
+	if !c.IgnoreSignals {
+		signal.Notify(c.osSignals, os.Kill, os.Interrupt, syscall.SIGTERM, syscall.SIGQUIT, syscall.SIGHUP)
+	}
 
 	var promptForPassword bool
 	// determine if they set the password flag but provided no value
@@ -95,17 +98,15 @@ func (c *CommandLine) Run() {
 	}
 
 	if err := c.Connect(""); err != nil {
-		fmt.Fprintf(os.Stderr,
-			"Failed to connect to %s\nPlease check your connection settings and ensure 'influxd' is running.\n",
+		return fmt.Errorf(
+			"Failed to connect to %s\nPlease check your connection settings and ensure 'influxd' is running.",
 			c.Client.Addr())
-		return
 	}
 
 	if c.Execute == "" && !c.Import {
 		token, err := c.DatabaseToken()
 		if err != nil {
-			fmt.Fprintf(os.Stderr, "Failed to check token: %s\n", err.Error())
-			return
+			return fmt.Errorf("Failed to check token: %s", err.Error())
 		}
 		if token == "" {
 			fmt.Printf(noTokenMsg)
@@ -125,15 +126,14 @@ func (c *CommandLine) Run() {
 		}
 
 		c.Line.Close()
-		os.Exit(0)
+		return nil
 	}
 
 	if c.Import {
 		path := net.JoinHostPort(c.Host, strconv.Itoa(c.Port))
 		u, e := client.ParseConnectionString(path, c.Ssl)
 		if e != nil {
-			fmt.Println(e)
-			return
+			return e
 		}
 
 		config := v8.NewConfig()
@@ -150,12 +150,12 @@ func (c *CommandLine) Run() {
 
 		i := v8.NewImporter(config)
 		if err := i.Import(); err != nil {
-			fmt.Printf("ERROR: %s\n", err)
+			err = fmt.Errorf("ERROR: %s\n", err)
 			c.Line.Close()
-			os.Exit(1)
+			return err
 		}
 		c.Line.Close()
-		os.Exit(0)
+		return nil
 	}
 
 	c.Version()
@@ -178,6 +178,7 @@ func (c *CommandLine) Run() {
 			close(c.Quit)
 		case <-c.Quit:
 			c.exit()
+			return nil
 		default:
 			l, e := c.Line.Prompt("> ")
 			if e == io.EOF {
@@ -824,6 +825,4 @@ func (c *CommandLine) exit() {
 	// release line resources
 	c.Line.Close()
 	c.Line = nil
-	// exit CLI
-	os.Exit(0)
 }
